move all custom hotkeys to files, fix something
