+++ config.h
@@ -6,8 +6,6 @@
 #define FIRMWARE_VERSION u8"xbJpL/wy5xo"
 #define RGB_MATRIX_STARTUP_SPD 60
 
-#undef TAPPING_TOGGLE
-#define TAPPING_TOGGLE 2
 #undef RGB_MATRIX_MAXIMUM_BRIGHTNESS
 #define RGB_MATRIX_MAXIMUM_BRIGHTNESS 255
