+++ rules.mk
@@ -8,3 +8,4 @@ RGBLIGHT_ENABLE = no
 RGB_MATRIX_ENABLE = yes
 CONSOLE_ENABLE = yes
 POINTING_DEVICE_ENABLE = no
+DYNAMIC_MACRO_ENABLE = no
